requirements.txt
amqp==5.0.9
asgiref==3.5.0
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.11
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-common-helpers==0.9.2
django-cors-headers==3.11.0
django-filter==21.1
django-redis==5.2.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
drf-yasg==1.20.0
flake8==4.0.1
flake8-bugbear==22.1.11
flake8-builtins==1.5.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
Markdown==3.3.6
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.3
prompt-toolkit==3.0.26
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.2
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.7.0
